Factor species changing out of wizmode (amalloy)
Species changing needs to be available on builds without wizmode for
this save compat thing to work.  This also does error handling in a
somewhat cleaner way.
